initial commit
sph based on o2d which is in turn based on a "Bullet Extra"
(http://bullet.googlecode.com/svn/trunk/Extras/sph/fluids/fluid_system.cpp)

neighbour hashing uses open adressing with linear probing (+1 intervals) and
stores the actual particle (not a reference) in the hash. the idea is to
improve locality of reference when iterating neighbour pairs, which is where
most of the complexity is by far.