rquad
a rasterizer for clipped quads. intended to be used for solid<->solid collision
detection. it's wrapped in a test suite, but it appears to have a number of bugs:
 - it makes out-of-bounds s/t coords (don't know if it can be avoided,
   currently prevented by clamping)
 - it sometimes makes gaps between triangles

it could be optimized to render spans instead of triangles (which is a bit
overkill)

it's based on mesa3d's software rasterizer (src/mesa/swrast/s_tritemp.h mostly)